# EDIT - Controllers changed private  ### 변경점 1. public에 있던 RDB, KV 컨트롤러가 private로 옮겨졌습니다. 각 스토리지에 접근하기 위해서는 Chain에 있는 wrapper를 이용하여 각 컨트롤러로 전달되어 처리됩니다.
2. 각 파일들을 수정하다 보니 hpp only로 구현되어있던 block_validator가
중복 선언으로 에러가 나서 cpp로 내용 구현을 옮겼고, 원래 스토리지들을
관리하려던 storage_manager의 역할을 Chain이 모두 하고있어서 해당 파일은
삭제하였습니다.
