+++ src/plugins/chain_plugin/include/chain.hpp
@@ -6,6 +6,7 @@
 #include "../structure/block.hpp"
 #include "kv_store.hpp"
 #include "rdb_controller.hpp"
+#include "unresolved_block_pool.hpp"
 
 #include <boost/program_options/variables_map.hpp>
 #include <memory>
@@ -17,9 +18,6 @@ using namespace std;
 
 class Chain {
 public:
-  unique_ptr<RdbController> rdb_controller;
-  unique_ptr<KvController> kv_controller;
-
   Chain(string_view dbms, string_view table_name, string_view db_user_id, string_view db_password);
   ~Chain();
 
@@ -29,8 +27,18 @@ public:
 
   void startup(nlohmann::json &genesis_state);
 
+  // RDB functions
   void insertBlockData(Block &first_block);
+
+  // KV functions
+  void saveWorld(world_type &world_info);
+  void saveChain(local_chain_type &chain_info);
+  void saveBackup(UnresolvedBlock &block_info);
+  string getValueByKey(DataType what, const string &base_keys);
+
 private:
   unique_ptr<class ChainImpl> impl;
+  unique_ptr<RdbController> rdb_controller;
+  unique_ptr<KvController> kv_controller;
 };
 } // namespace gruut
